diagnostics: Follow DECL_ORIGIN in lhd_print_error_function [PR102061]
Currently, if a warning references a cloned function, the name of the cloned
function will be emitted in the "In function 'xyz'" part of the diagnostic,
which users aren't supposed to see. This patch follows the DECL_ORIGIN link
to get the name of the original function, so the internal compiler details
aren't exposed.

gcc/ChangeLog:
	PR diagnostics/102061
	* langhooks.cc (lhd_print_error_function): Follow DECL_ORIGIN
	links.
	* gcc.dg/pr102061.c: New testcase.