[PR119765] testsuite: adjust amd64-abi-9.c to check both ms and sysv ABIs
This test was failing because it was checking that eax was being cleared. For
sysv abi, eax contains the number of XMM registers used in the call, but msabi
just passes the float arguments twice, both in xmm and general purpose registers.

This patch adds tests for both sysv and msabi functions being called, and adds
some more tests that check for for argument being placed in the correct
register.

Tested on x86_64-linux-gnu and x86_64-w64-mingw32

gcc/testsuite:
	PR testsuite/119765
	* gcc.target/i386/amd64-abi-9.c: Add sysv attribute to called function
	Add msabi attribute function
	Add test checking parameter is placed in correct register

Co-Authored-By: NightStrike <nightstrike@gmail.com>