Fix errno handling for sin and cos [PR80042]
POSIX says that sin and cos should set errno to EDOM when infinity is passed to
them. Make sure this is accounted for in builtins.def.

When sin/cos are called with values that set errno (like INFINITY), GCC was
incorrectly optimizing the code by assuming these functions never modify errno.
This caused incorrect behavior at -O2 where errno reads could be moved relative
to the sin/cos calls.

pass_cse_sincos also needs to be disabled when errno matters.
GCC internally optimizes separate sin() and cos() calls to a call to
the cexpi builtin, and then that builtin "evaluates" to sincos later (for
targets which have it), otherwise cexpi is used (which does not set errno).
Guarding this transformation behind -fno-math-errno works, but it
unnecessarily pessimizes code into separate sin and cos calls.

Fixing this properly might require expanding the sincos builtin to be able to emit
separate sin and cos calls for targets which don't have sincos, and then
optimizing to sincos for the case of -fmath-errno. For now, we preserve the
individual sin/cos calls when errno matters to maintain correct behavior.

gcc/ChangeLog:

	PR middle-end/80042
	* builtins.def (BUILT_IN_SIN, BUILT_IN_SINF, BUILT_IN_SINL): Change
	from ATTR_MATHFN_FPROUNDING to ATTR_MATHFN_FPROUNDING_ERRNO.
	(BUILT_IN_COS, BUILT_IN_COSF, BUILT_IN_COSL): Likewise.
	* tree-ssa-math-opts.cc (pass_cse_sincos::execute): Don't optimize
	sin/cos to cexpi when flag_errno_math is enabled, since cexpi
	doesn't set errno.